A new lockless priority queue structure
Each priority queue is now two queues:
- a lockless linked list
- a standard STL std::priority_queue

The lockless linked list is accessed by all other threads as a way to push
events onto the list. The lockless linked list uses atomics. The executing
thread pulls from the lockless linked list and pushes into the priority queue.
The executing thread is now the only thread that accesses the priority queue.
Spinlocks (or any locks) are not used in the design.

I found that significant lock contention was occuring when events per crossing
executer thread boundaries. The new update removes all of this contention.