Implemented executer mapping API
The simulator now takes an executer mapping object so that the user
can specify how components get mapped to executers. The API is specified
as a simple Mapper class. Subclasses to Mapper only have to override a
single function that take in the num of executers and the component then
return the executer ID that should be used. If a single executer is used,
the user doesn't need to be concerned about this.

The util/ direction now contains to Mapper implementations:
RoundRobin - a perfectly load-balancing sweeping mapping (same as previous)
Random - a random component to executer mapping